Move file generation into setup.py.
Moves the code for creating the generated pyi files out of release into
build_utils and calls it from setup.py.

This change allows setup.py to do a complete pytype installation independent of
the release script, which (1) is useful for testing and (2) will enable
https://github.com/python/typeshed/blob/master/requirements-tests-py2.txt to go
back to testing with pytype at HEAD.

Also bumps the version number so that I can do a PyPI release and check that
everything is working.

Tested by running the release script at an old version number (as before) and
with `pip install . -U --user`.

PiperOrigin-RevId: 204563486
