+++ pytype/__version__.py
@@ -1,2 +1,2 @@
 # pylint: skip-file
-__version__ = '2018.6.19'
+__version__ = '2018.7.13'
